FAC-811: refactor - Move Sort enum from common to core module
